Fix test time out on feature-propagation-to-auxiliary-context.html
The test is flaky because the number of subtests run depends on how
many features are available in feature policy. This number grows through
out time and eventually causes a timeout occasionally.

The solution is to run a single test that includes all features, instead
of running a subtest per feature.

Bug: 1076410
Change-Id: I9085768d3e000892cef94d661552f4cb518e08ee